Input bug probably fixed. A null event was returned so ``` if(os_event(events, key, key_down)) {
}
```
Was still registered as valid. Since a null event != 0

I just removed something silly I was doing and that fixed it.
I don't remember why I did it in the first place. Check the
changes in os_glfw.c for that.

Also, cleared input mv when disabling camera. This stops the camera from
retaining movement info on resuming.